[refactor] module.scss 컨벤션에 맞춰서 파일 이름 수정
